+++ examples/realworld/src/App.js
@@ -1,11 +1,11 @@
 import React from 'react'
 import { component, createSub } from 'framework-x'
-import { Profile } from './components/Profile'
+import { Profile } from './profile/views'
 import { routeIds } from './routes'
 import { getRouteId } from './routes/selectors'
 import { Header } from './components/Header'
 import { Article } from './articles/views'
-import { Editor } from './components/Editor'
+import { Editor } from './editor/views'
 import { Home } from './components/Home'
 import { Login } from './auth/views'
 import { Register } from './auth/views'
