+++ examples/realworld/src/profile/views.js
@@ -2,11 +2,11 @@ import React from 'react'
 import * as R from 'ramda'
 import { component, createSub, derive } from 'framework-x'
 import { evt } from '../eventTypes'
-import { getProfileUser, isViewingSelf } from '../profile/selectors'
+import { getProfileUser, isViewingSelf } from './selectors'
 import { routeIds } from '../routes'
 import { getRouteId } from '../routes/selectors'
-import { ArticleList } from './ArticleList'
-import { Link, NavLink } from './Link'
+import { ArticleList } from '../components/ArticleList'
+import { Link, NavLink } from '../components/Link'
 import { dispatch } from '../store'
